Закончил

+++ core/location-service/src/main/java/ru/practicum/service/LocationServiceImpl.java
@@ -6,6 +6,8 @@ import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.clients.event.EventClient;
+import ru.practicum.core.error.exception.ConflictDataException;
 import ru.practicum.core.error.exception.NotFoundException;
 import ru.practicum.core.util.PagingUtil;
 import ru.practicum.dto.location.LocationDto;
@@ -26,6 +28,7 @@ import java.util.Optional;
 public class LocationServiceImpl implements LocationService {
     LocationRepository locationRepository;
     LocationMapper locationMapper;
+    EventClient eventClient;
 
     @Override
     public List<LocationDto> getLocations(Integer from, Integer size) {
@@ -47,9 +50,10 @@ public class LocationServiceImpl implements LocationService {
     @Override
     @Transactional
     public LocationDto addLocation(NewLocationDto newLocationDto) {
-        Location location = locationRepository.save(locationMapper.toLocation(newLocationDto));
+        Location location = locationMapper.toLocation(newLocationDto);
+        Location locationSaved = locationRepository.save(locationMapper.toLocation(newLocationDto));
         log.info("Location is created: {}", location);
-        return locationMapper.toDto(location);
+        return locationMapper.toDto(locationSaved);
     }
 
     @Override
@@ -66,27 +70,50 @@ public class LocationServiceImpl implements LocationService {
     @Override
     @Transactional
     public void delete(Long locationId) {
+
+        if (eventClient.get(locationId)) {
+            throw new ConflictDataException("Есть связанные с этим местоположением события");
+        }
         locationRepository.deleteById(locationId);
         log.info("Location deleted with id: {}", locationId);
     }
+
     @Transactional
     @Override
     public LocationDto findLocationBy(NewLocationDto newLocationDto) {
-        log.info("start findLocationById");
+        log.info("start findLocationBy newLocationDto {}", newLocationDto);
+        if (newLocationDto.getLat() == 0 && newLocationDto.getLon() == 0) {
+            newLocationDto.setLat(0.0);
+            newLocationDto.setLon(0.0);
+        }
 
-        Optional<Location> location = locationRepository.findByLatAndLon(newLocationDto.getLat(), newLocationDto.getLon());
+        if (newLocationDto == null) {
+            log.info("newLocationDto is null, условие сработало");
+            return null;
+        }
+        Optional<Location> location = locationRepository.findLocationByLatAndLon(newLocationDto.getLat(), newLocationDto.getLon());
         Location newLocation;
-        if (!location.isPresent()) {
+
+
+        if (location.isEmpty()) {
             log.info("местоположение не найдено, сохраняем");
             newLocation = locationRepository.save(locationMapper.toLocation(newLocationDto));
         } else {
+            log.info("местоположение найдено, достаем из БД id={}", location.get().getId());
             newLocation = location.get();
         }
         log.info("findLocationBy result location = {}", newLocation);
         return locationMapper.toDto(newLocation);
     }
 
-
-    // return locationMapper.toLocationDto(newLocationDto) == null ? null : locationRepository.findByLatAndLon(newLocationDto.getLat(), newLocationDto.getLon())
-    //       locationMapper.toDto(locationRepository.save(locationMapper.toLocation(newLocationDto))));
+    @Override
+    public List<LocationDto> getLocationByRadius(Double lat, Double lon, Double radius) {
+        log.info("Поиск местоположений по радиусу и координатам, lat, lon, radius");
+        List<LocationDto> locationDtos = locationRepository.findByRadius(lat, lon, radius)
+                .stream()
+                .map(locationMapper::toDto)
+                .toList();
+        log.info("Нашли список локаций: ", locationDtos);
+        return locationDtos;
+    }
 }
